Export `MockedProviderProps` and `MockedProviderState`
These were available in `react-apollo@2`, and there isn't
any reason not to expose them.

Fixes #3336.
